Refactor : 주변 명소 스팟 상세 코드 정리
