+++ src/page/main/Main.tsx
@@ -1,40 +1,60 @@
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Map } from "react-kakao-maps-sdk";
 import Input from "../../components/common/input/Input";
 import Navigation from "../../components/common/menu/Navigation";
 import BlossomMarker from "../../components/common/marker/BlossomMarker";
-import { SpotView } from "../../components/main/SpotView";
 import { MainBookMarkBtn } from "../../components/main/Btn/MainBookMarkBtn";
-import { getCurrentLocation } from "../../utils/mapLocation/getCurrentLocation";
 import { useAppDispatch, useAppSelector } from "../../store/hooks";
 import { useWatchLocation } from "../../hook/mapLocation/useWatchLocation";
 import UserMarker from "../../components/common/marker/UserMarker";  
 import { getMarker } from "../../store/features/main/marker/markerSlice";
 import { hide } from "../../store/features/main/map/tourClick";
 import { spotHide, spotShow } from "../../store/features/main/map/spotClick";
-import SpotIntro from "../../components/main/SpotIntro";
 import { getPlace } from "../../store/features/main/spotInfo/InfoPlace";
 import { useGetPostion } from "../../hook/mapLocation/useGetPostion";
 import FlowerTag from "../../components/common/tag/FlowerTag";
 import { Link } from "react-router-dom";
-import TopBar from "../../components/common/menu/TopBar";
+import { LocationBtn } from "../../components/main/Btn/LocationBtn";
+import Tour from "../../components/main/Tour/Tour";
+import { Info } from "../../components/main/Info/Info";
+import { MainMarker } from "../../store/@types/main/marker/MarkerType";
 
 export default function Main() {
 
   const dispatch = useAppDispatch();
   const { userInfo } = useAppSelector((state) => state.auth);
+  
   // 북마크
   const [bookMark, setBookMark] = useState(false);
-  // 마커
+
+  // 등록된 모든 마커
   const { markers } = useAppSelector((state) => state.mainMarker);
   useEffect(() => {
     if (!userInfo) return;
     dispatch(getMarker(userInfo.token));
   }, [bookMark, userInfo]);
+
   // 지도 초기위치 설정, 포지션 가져오기
   const {state,setState,getPostion} = useGetPostion();
+
   // 마커 클릭
   const [clickMarker,setClickMarker] = useState(0);
+  const markerHanlder = (e : MainMarker)=>{
+    if (!userInfo) return;
+    // info 데이터
+    setClickMarker(e.placeId);
+    setState({
+      center : {
+        lat: e.x, 
+        lng: e.y
+      },
+      isPanto : true
+    })
+    dispatch(
+      getPlace({ token: userInfo.token, placeId: e.placeId }),
+    );
+  }
+
   // 중심좌표 변경
   const onCenterChanged = (map : kakao.maps.Map)=>{
     const latlng = map.getCenter();
@@ -46,8 +66,11 @@ export default function Main() {
       isPanto: false,
     });
   }
+
   // 현재위치 Watch
   const { location } = useWatchLocation();
+
+  // 검색기능
   const [search,setSearch] = useState(false);
   const searchHandler = ()=>{
     setSearch(true);
@@ -71,17 +94,6 @@ export default function Main() {
               setBookMark={setBookMark} 
             />
           }
-          
-          {/* 임시 로그아웃 버튼 */}
-          {
-            !search &&
-            <button
-              onClick={() => dispatch(logout())}
-              className="m-2 bg-white p-2"
-            >
-              로그아웃
-            </button>
-          }
         </div>
 
         {/* 맵 */}
@@ -91,9 +103,7 @@ export default function Main() {
           style={{ width: "100%", height: "100%" }}
           level={5}
           onDragStart={() => {
-            dispatch(hide());
-            dispatch(spotHide());
-            setClickMarker(0);
+
           }}
           onCenterChanged={onCenterChanged}
         >
@@ -103,15 +113,7 @@ export default function Main() {
             <BlossomMarker
               key={i + e.x + e.y}
               position={{ lat: e.x, lng: e.y }}
-              onClick={() => {
-                if (!userInfo) return;
-                // info 데이터
-                dispatch(hide());
-                dispatch(spotShow());
-                dispatch(
-                  getPlace({ token: userInfo.token, placeId: e.placeId }),
-                );
-              }}
+              onClick={() => markerHanlder(e)}
             />
           ))}
 
@@ -121,8 +123,6 @@ export default function Main() {
             onClick={() => {
               if (!userInfo) return;
               // info 데이터
-              dispatch(hide());
-              dispatch(spotShow());
               setClickMarker(1);
               setState({
                 center : {
@@ -130,12 +130,15 @@ export default function Main() {
                   lng: 126.69745010724675
                 },
                 isPanto : true
-              })
+              });
+              dispatch(
+                getPlace({ token: userInfo.token, placeId: 1 }),
+              );
             }}
             isClicked={clickMarker === 1}
           />
 
-          {/* 유저 */}
+          {/* 유저 마커 */}
           {
             location && (
               <UserMarker
@@ -148,16 +151,29 @@ export default function Main() {
         </Map>
 
         {/* 마커 관련 명소 */}
-        <SpotIntro 
-          getPostion={getPostion} 
-        />
+        {/* <SpotIntro getPostion={getPostion}/> */}
 
+        {/* 마커 클릭 관련 따로 만들기 */}
+        
+        <div className={`absolute h-[calc(100vh-70px)] bottom-[70px] z-[60] w-full transition-transform duration-300`}>
+            <div className={`absolute bottom-full left-5 z-20 mb-5 transition-transform duration-500}`}>
+              <LocationBtn onClick={getPostion} />
+            </div>
+            <Tour/>
+        </div>
+       
+
+        {/* <div className={`absolute bottom-[70px] z-[60] w-full transition-transform duration-300`}>
+          <div className={`absolute bottom-full left-5 z-20 mb-5 transition-transform duration-500}`}>
+            <LocationBtn onClick={getPostion} />
+          </div>
+          <Info />
+        </div> */}
+        
         {/* 네비게이션바 */}
         <Navigation />
       </main>
-
-      <SpotView />
-
+      
       {
         search &&
         <div className="bg-white absolute top-0 left-0 w-full h-full z-50 pt-[95px] tracking-custom">
@@ -184,12 +200,3 @@ export default function Main() {
     </>
   );
 }
-
-/*
-const [state, setState] = useState({
-  // 지도의 초기 위치
-  center: { lat: 13, lng: 14 },
-  // 지도 위치 변경시 panto를 이용할지에 대해서 정의
-  isPanto: false,
-});
-*/
