+++ src/store/@types/main/marker/MarkerType.ts
@@ -1,10 +1,10 @@
-type MainMarker = {
+export interface MainMarker {
     placeId: number;
     x: number;
     y: number;
 }
 
-export interface initialStateType {
+export interface MainMarkerType {
     loading : boolean;
     markers : MainMarker[];
     error : boolean;
