+++ src/store/@types/main/tourList/TourPlaceType.ts
@@ -1,0 +1,18 @@
+export interface TourPlaceType {
+    loading : boolean
+    data : {
+        placeDetails: PlaceDetail[];
+        hasNext: boolean;
+    }
+    error : null | string
+}
+
+export interface PlaceDetail {
+    placeId: number;
+    name: string;
+    addressDetail: string;
+    placeImage: string;
+    bookmarkStatus: boolean;
+    placeStatus: string;
+}
+
