+++ src/store/features/main/marker/markerSlice.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import axios from "axios";
-import { initialStateType } from "../../../@types/main/marker/MarkerType";
+import { MainMarkerType } from "../../../@types/main/marker/MarkerType";
 
 export const getMarker = createAsyncThunk("get/marker",async(token : string)=>{
 
@@ -16,21 +16,8 @@ export const getMarker = createAsyncThunk("get/marker",async(token : string)=>{
 
 });
 
-export const getBookMarker = createAsyncThunk("get/bookmarker",async(token : string)=>{
-
-    const response = await axios.get('/api/v1/place/all?tag=벚꽃',{
-        headers : {
-            Authorization : token
-        }
-    });
-    const {data} = response;
-    console.log(data);
-    return data;
-
-});
-
-const initialState : initialStateType = {
-    loading : true,
+const initialState : MainMarkerType = {
+    loading : false,
     markers : [],
     error : false
 };
@@ -42,21 +29,14 @@ const mainMarker = createSlice({
     extraReducers : (builder) => {
         builder
         .addCase(getMarker.pending,(state)=>{
-            state.loading = false;
+            state.loading = true;
         })
         .addCase(getMarker.fulfilled,(state,action)=>{
+            state.loading = false;
             state.markers = action.payload;
         })
         .addCase(getMarker.rejected,(state)=>{
-            state.error = true;
-        })
-        .addCase(getBookMarker.pending,(state)=>{
             state.loading = false;
-        })
-        .addCase(getBookMarker.fulfilled,(state,action)=>{
-            state.markers = action.payload;
-        })
-        .addCase(getBookMarker.rejected,(state)=>{
             state.error = true;
         })
     }
