+++ src/store/store.ts
@@ -6,13 +6,10 @@ import { buildGetDefaultMiddleware } from "@reduxjs/toolkit/dist/getDefaultMiddl
 import authSlice from "./features/auth/authSlice";
 import mainMarker from "./features/main/marker/markerSlice";
 import spotViewReducer from "./features/main/spotView/slice";
-import spotInfoY from "./features/main/spotInfo/ySlice";
 import spotInfoOpen from "./features/main/spotInfo/InfoSlice";
 import tourOpen from "./features/main/tourList/openSlice";
 import InfoPlace from "./features/main/spotInfo/InfoPlace";
 import tourPlace from "./features/main/tourList/tourPlace";
-import tourClick from "./features/main/map/tourClick";
-import spotClick from "./features/main/map/spotClick";
 import errorModal from "./features/errorModal/modalSlice";
 
 const persistConfig = {
@@ -24,13 +21,10 @@ const persistConfig = {
 const rootReducer = combineReducers({
   auth: authSlice,
   spotView: spotViewReducer,
-  spotInfoY,
   spotInfoOpen,
   tourOpen,
   mainMarker,
   InfoPlace,
-  tourClick,
-  spotClick,
   tourPlace, // 주변 명소
   errorModal,
 });
